html/semantics/embedded-content/the-iframe-element/iframe-loading-lazy-nav-link-click.html

<!DOCTYPE html>
<title>Navigating iframe loading='lazy' before it is loaded: link click</title>
<iframe name="iframe" src="support/blank.htm?src" loading="lazy"></iframe>
<a href="support/blank.htm?nav" target="iframe"></a>
<script>
const iframe = document.querySelector('iframe');
const a = document.querySelector('a');
a.click();
</script>
<!-- Loading testharness.js here is intentional to reproduce a bug in WebKit. -->
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<script>
setup({single_test: true});
onload = () => {
  step_timeout(() => {
    assert_equals(iframe.contentWindow.location.href, new URL("support/blank.htm?nav", location.href).href);
    done();
  }, 1000);
};
</script>
<!-- Delay the load event for the same URLs as the potential lazy iframe navigations
     to reduce unstable test results. -->
<iframe src="support/blank.htm?src"></iframe>
<iframe src="support/blank.htm?nav"></iframe>
